+++ simulations/geometryXYVxVy/spline_definitions_xyvxvy.hpp
@@ -48,16 +48,14 @@ using SplineInterpolatorY = SplineInterpolator<
         BSplinesY,
         GridY,
         ExtrapolationRule::Periodic,
-        SplineYClosure,
-        SplineYClosure>;
+        SplineBoundaryClosures<SplineYClosure, SplineYClosure>>;
 
 using SplineInterpolatorVy = SplineInterpolator<
         Kokkos::DefaultExecutionSpace,
         BSplinesVy,
         GridVy,
         ExtrapolationRule::Constant_Constant,
-        SplineVyClosure,
-        SplineVyClosure>;
+        SplineBoundaryClosures<SplineVyClosure, SplineVyClosure>>;
 
 using IdxRangeBSY = IdxRange<BSplinesY>;
 using IdxRangeBSXY = IdxRange<BSplinesX, BSplinesY>;